refactor(global): change whole projects structure to work with typescript
Change projects language struct to typescript integration to properly work with node and ts
environments. Make changes to the config files for the necessary typescript directory
implementations.

BREAKING CHANGE: Whole project runs on the typescript dependant environment

#29
