+++ .storybook/preview.ts
@@ -3,7 +3,6 @@ import 'remixicon/fonts/remixicon.css'
 import colors from 'tailwindcss/colors'
 import { addons } from '@storybook/preview-api'
 import type { Preview } from '@storybook/vue3'
-
 const preview: Preview = {
   parameters: {
     actions: { argTypesRegex: '^on[A-Z].*' },
@@ -57,89 +56,6 @@ const preview: Preview = {
 }
 
 export default preview
-
-// export const parameters = {
-//   actions: { argTypesRegex: "^on[A-Z].*" },
-//   backgrounds: {
-//     default: "light",
-//     dark: 'dark',
-//     values: [
-//       {
-//         name: "white",
-//         value: "#FFFFFF",
-//       },
-//       {
-//         name: "black",
-//         value: "#000000",
-//       },
-//       {
-//         name: "dark",
-//         value: colors.gray[800],
-//       },
-//       {
-//         name: "light",
-//         value: colors.gray[100],
-//       },
-//       {
-//         name: "warning",
-//         value: colors.yellow[500],
-//       },
-//       {
-//         name: "danger",
-//         value: colors.red[500],
-//       },
-//       {
-//         name: "primary",
-//         value: colors.blue[500],
-//       },
-//       {
-//         name: "success",
-//         value: colors.green[500],
-//       },
-//       {
-//         name: "info",
-//         value: colors.cyan[500],
-//       },
-//       {
-//         name: "secondary",
-//         value: colors.gray[500],
-//       },
-//     ],
-//   },
-// };
-// export const argTypes = {
-//   variant: { control: "select", options: ["solid", "outline", "text", "link"] },
-//   color: {
-//     control: "select",
-//     options: ["primary", "secondary", "success", "danger", "warning", "info"],
-//   },
-//   filter: {
-//     control: "select",
-//     options: ["lighten", "darken", "none"],
-//   },
-//   block: {
-//     control: "boolean",
-//   },
-//   rounded: {
-//     control: "select",
-//     options: [true, false, "full"],
-//   },
-//   disabled: {
-//     control: "boolean",
-//   },
-//   prepend: {
-//     control: "select",
-//     options: ["home", "add-circle", "close", "none"],
-//   },
-//   append: {
-//     control: "select",
-//     options: ["home", "add-circle", "close", "none"],
-//   },
-//   size: {
-//     control: { type: "select" },
-//     options: ["sm", "md", "lg"],
-//   },
-// };
 
 // get an instance to the communication channel for the manager and preview
 const channel = addons.getChannel()
