+++ src/components/Button/LuiButton.stories.ts
@@ -1,4 +1,6 @@
 import LuiButton from './LuiButton.vue'
+import type { Meta, StoryObj } from '@storybook/vue3'
+
 import {
   variant,
   color,
@@ -8,12 +10,11 @@ import {
   filter,
   disabled
 } from '../../../.storybook/global-story-argtypes'
+import { ButtonSize } from './button-types'
 
-// More on default export: https://storybook.js.org/docs/vue/writing-stories/introduction#default-export
-export default {
+const meta: Meta<typeof LuiButton> = {
   title: 'LUI/Button',
   component: LuiButton,
-  // More on argTypes: https://storybook.js.org/docs/vue/api/argtypes
   argTypes: {
     variant,
     color,
@@ -31,26 +32,27 @@ export default {
     }
   }
 }
+export default meta
+type Story = StoryObj<typeof LuiButton>
 
-// More on component templates: https://storybook.js.org/docs/vue/writing-stories/introduction#using-args
-const Template = (args) => ({
-  // Components used in your story `template` are defined in the `components` object
-  components: { LuiButton },
-  setup() {
-    return { args }
-  },
-  // And then the `args` are bound to your component with `v-bind="args"`
-  template: `<lui-button v-bind="args">My Button</lui-button>`
-})
-export const Default = Template.bind({})
-Default.args = { rounded: true }
-export const Rounded = (args) => ({
-  components: { LuiButton },
-  setup() {
-    const roundedValues = [false, true, 'full']
-    return { args, roundedValues }
-  },
-  template: `
+export const Default: Story = {
+  render: (args) => ({
+    components: { LuiButton },
+    setup() {
+      return { args }
+    },
+    template: `<lui-button v-bind="args">My Button</lui-button>`
+  })
+}
+
+export const Rounded: Story = {
+  render: (args) => ({
+    components: { LuiButton },
+    setup() {
+      const roundedValues = [false, true, 'full']
+      return { args, roundedValues }
+    },
+    template: `
   <div class="flex space-x-4">
       <div class="flex flex-col" v-for="(value,index) in roundedValues" :key="index">
       <lui-button v-bind="args" :rounded="roundedValues[index]">
@@ -58,27 +60,31 @@ export const Rounded = (args) => ({
     </lui-button>
       </div>
   </div>`
-})
-export const Variants = (args) => ({
-  components: { LuiButton },
-  setup() {
-    return { args }
-  },
-  template: `
+  })
+}
+export const Variants: Story = {
+  render: (args) => ({
+    components: { LuiButton },
+    setup() {
+      return { args }
+    },
+    template: `
   <div class="space-x-2">
     <lui-button v-bind="args" variant="solid">Solid</lui-button>
     <lui-button v-bind="args" variant="outline">Outline</lui-button>
     <lui-button v-bind="args" variant="link">Link</lui-button>
     <lui-button v-bind="args" variant="text">Text</lui-button>
   </div>`
-})
-export const Colors = (args) => ({
-  components: { LuiButton },
-  setup() {
-    const colors = ['primary', 'secondary', 'info', 'success', 'warning', 'danger']
-    return { args, colors }
-  },
-  template: `
+  })
+}
+export const Colors: Story = {
+  render: (args) => ({
+    components: { LuiButton },
+    setup() {
+      const colors = ['primary', 'secondary', 'info', 'success', 'warning', 'danger']
+      return { args, colors }
+    },
+    template: `
   <div class="space-x-2">
     <lui-button v-for="color in colors" :key="color" v-bind="args" :color="color">
       {{color}}
@@ -96,26 +102,19 @@ export const Colors = (args) => ({
   </div>
   
   `
-})
-export const Sizes = (args) => ({
-  components: { LuiButton },
-  setup() {
-    const sizes = ['xs', 'sm', 'md', 'lg', 'xl']
-    // const iconSize = computed(() =>
-    //   props.size === "xs"
-    //     ? "12"
-    //     : props.size === "sm"
-    //     ? "16"
-    //     : props.size === "xl"
-    //     ? "24"
-    //     : "20"
-    // );
-    function iconSize(size) {
-      return size === 'xs' ? '12' : size === 'sm' ? '16' : size === 'xl' ? '24' : '20'
-    }
-    return { args, sizes, iconSize }
-  },
-  template: `<div class="flex items-center space-x-4">
+  })
+}
+export const Sizes: Story = {
+  render: (args) => ({
+    components: { LuiButton },
+    setup() {
+      const sizes = ['xs', 'sm', 'md', 'lg', 'xl']
+      function iconSize(size: ButtonSize) {
+        return size === 'xs' ? '12' : size === 'sm' ? '16' : size === 'xl' ? '24' : '20'
+      }
+      return { args, sizes, iconSize }
+    },
+    template: `<div class="flex items-center space-x-4">
     <div v-for="size in sizes" :key="size">
       <lui-button v-bind="args" :size="size">
         <span>LuiButton</span>
@@ -125,19 +124,20 @@ export const Sizes = (args) => ({
       </lui-button>
     </div>
   </div>`
-})
-
-export const IconButton = (args) => ({
-  components: { LuiButton },
-  setup() {
-    const sizes = ['xs', 'sm', 'md', 'lg', 'xl']
-    const roundedValues = [false, true, 'full']
-    function iconSize(size) {
-      return size === 'xs' ? '12' : size === 'sm' ? '16' : size === 'xl' ? '24' : '20'
-    }
-    return { args, sizes, roundedValues, iconSize }
-  },
-  template: `
+  })
+}
+export const IconButton: Story = {
+  render: (args) => ({
+    components: { LuiButton },
+    setup() {
+      const sizes = ['xs', 'sm', 'md', 'lg', 'xl']
+      const roundedValues = [false, true, 'full']
+      function iconSize(size: ButtonSize) {
+        return size === 'xs' ? '12' : size === 'sm' ? '16' : size === 'xl' ? '24' : '20'
+      }
+      return { args, sizes, roundedValues, iconSize }
+    },
+    template: `
   <div class="space-y-6">
     <div v-bind="args" class="flex items-center space-x-6" v-for="value in roundedValues" :key="value">
       <div class="flex" v-for="size in sizes" :key="size">
@@ -149,4 +149,5 @@ export const IconButton = (args) => ({
       </div>
     </div>
   </div>`
-})
+  })
+}
