+++ src/components/Button/button-types.ts
@@ -1,3 +1,3 @@
 export type ButtonTag = 'button' | 'div'
-// export type ButtonSize = "sm" | "md" | "lg";
+export type ButtonSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl'
 // export type loaderPosition = "right" | "left";
