+++ src/components/Input/input-types.ts
@@ -1,3 +1,4 @@
 export type Clear = false | true
 export type ModelValue = string | number
 export type Description = string | null
+export type IconSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl'
